Added DICOM reader that allows to manually select which configuration should be used.
This is not a fix for the complete problem.
But it allows to circumvent for testing purposes the
auto selecting policy "take configuration with the least number
of images", which causes the problem here as e.g. the configuration
"Instance Number, consecutive" generates 4 correct images for the
data of task T27483.