Add a separate action to auto-crop LabelSetImages
Handling of LabelSetImages is so quirky, that no functionality of
the classic auto-crop action or image filters in general can be reused.

As we do not want to invest more work than necessary in the ill-formed
LabelSetImage data type, I encapsulated/implemented the whole operation
right into the action class, instead of providing a "public" filter.