Remove everything related to subprojects/labels
This happened in preparation of the dashboard script rewrite to unravel
things a bit. We planned to get rid of the current subprojects anyway,
as they were not maintained for years and do not actually have any
surplus value on the dashboard.

We may consider reintroducing subprojects based on a different concept
like MITK extensions.