rewrote some sections and added new image
Summary:
Some sections needed to be changed due to plugin and gui changes.
Note: I am not entirely sure about the function "All Components". That
is why I spared it out on the screenshot but that could be changed if
necessary.

Fixes T27107

Test Plan: f1 or webpage

Reviewers: O3 MITK Reviewer Group II, gaoh

Reviewed By: O3 MITK Reviewer Group II, gaoh

Subscribers: gaoh

Maniphest Tasks: T27107

Differential Revision: https://phabricator.mitk.org/D377