Make Git a required dependency
We already assumed Git to be a required dependency in most possible
configurations except the very minimum ones. As we also want to
be able to use Git more often in ExternalProject_Add() calls I did
the final step of moving the find_package(Git) call to a very early
stage of the MITK (superbuild) script.