BREAKING ContourModelUtils::FillSliceInSlice
Why: Foreground detection was hard coded to 1<foreground.
This was ok for the use so far (FillContourInSlice), because there the
foreground value was 255. But was problematic for general usage.
Changed now the check to "fillForegroundThreshold<=foreground" and
the default of filForgroundThreshold to 1.

How to adapt: You only have to adapte if values of 1 are still background in
your case. Then you have to pass a fitting fillForegroundThreshold for your case.

+ Improved constness
+ Removed obsolete InternalWritePreviewOnWorkingImage
+ Removed unnecesary member CorrectorTool2D:m_WorkingSlices (reduces internal states)
+ Fixed T28225