BREAKING!!!
What:
- removed GetCheckPrecision() from mitk::NodePredicateGeometry
and mitk::NodePredicateSubGeometry.
- removed constant NODE_PREDICATE_GEOMETRY_DEFAULT_CHECK_PRECISION

Why:
The equality functions of geometries (equal and isSubGeometry) do support
coordinate precision and direction precision. This feature was not used
by the predicates so far. But it is needed to allow the predicates to be
a bit less pedantic with differences in coordinate comparisons. These
kind of differences happen with real world data, but have no practical
impact for normal medical imaging, thus should not impact UX.

How to adapt:
- GetCheckPrecision: Use GetCheckCoordinatePrecision or
  GetCheckDirectionPrecision instead, depending of which aspect you need.
- NODE_PREDICATE_GEOMETRY_DEFAULT_CHECK_PRECISION:
  Use NODE_PREDICATE_GEOMETRY_DEFAULT_CHECK_COORDINATE_PRECISION or
  NODE_PREDICATE_GEOMETRY_DEFAULT_CHECK_DIRECTION_PRECISION instead,
  depending of which aspect you need.

Additionally:
+ Fixed T28356