Rename vector of data nodes
This member variable was newly introduced for the "SetSelected" mode.
However, the name might be confusing since it's similar to the
"GetRelevantNodes"-function.
There is a difference in both sets of data nodes as the
"GetRelevantNodes"-function returns a filtered subset of the data storage,
whereas the new vector of "Data Nodes for LevelWindow" is a set of relevant
nodes that have been further filtered (e.g. visibility, rendering mode).