Modify test to retrieve the used levelwindow property
The data storage accesses the data nodes in a non-deterministic way.
In order to test the "imageForLevelWindow" / "levelwindow" properties,
the "levelwindow" property of all three nodes needs to be tested.
Only one node has the "imageForLevelWindow" property set and this node's
"levelwindow" property is used by the level window manager.