Send active render window
Summary:
Remove mouse event from render window

Listen to and react on a 'FocusChangedEvent'

This event is sent by the RenderingManager and will be used to set the
active render window of a multi widget.
The event is sent if a 'MousePressEvent' is processed.

React on an active render window part

In order to react on an already active render window part,
if the rendering manager plugin view was opened after the render window part,
'RenderWindowPartActivated' needs to be called during UI creation.

Test Plan:
manual
1. open MxNMultiWidget
2. create multiple render windows
3. click on different render windows and see how the border color changes
4. open the "RenderWindowManager"
5. open StdMultiWidget
6. click on differet render windows and see how the name in the combobox changes

Reviewers: kleina, O3 MITK Reviewer Group II, kislinsk

Reviewed By: O3 MITK Reviewer Group II, kislinsk

Subscribers: kislinsk, kalali, kleina

Maniphest Tasks: T26492

Differential Revision: https://phabricator.mitk.org/D553