Always expand all tree nodes
Since the decision a few years ago that a modified property should also
modify the property list it is part of, the property tree view collapses
all nested items every time a property value is changed inside the tree
view itself, since it indirectly modifies the whole property list which
in turn must trigger a model reset.

There is no real feasible solution to adapt to this behavior so we always
do the opposite now and expand everything which results in a more
consistent UX that minimizes fights against the users intent.

For a better overview, users can use the filter widget.