fix image masking issues for images with multiple timesteps
Summary:
Masking images with the segentation utilities led to an exception when having an image with multiple timesteps and choosing a custom value as background pixel value.
Also, in images with timesteps, masking was not possible due to disabled widgets with the message "Different image sizes cannot be masked".
Therefore, I changed the access by ITK to work with the 4D case (fixes the custom background issue), and only check if the image size is equal in x, y and z dimension (fixes the disabled widgets).
Fixes T28795

Test Plan: Manual, see T28795 for test cases

Reviewers: kalali, O1 MITK Reviewer Group I, kahl

Subscribers: floca

Maniphest Tasks: T28795

Differential Revision: https://phabricator.mitk.org/D566