Do not show dialog on new segmentation creation
In order to clean the process of adding a new segmentation node,
the new segmentation dialog was removed and a new label set image
is created without any input prompt.
A new layer and a first label is automatically added to the label
set image but no segmentation color, organ or label name needs to
be provided.
The segmentation node is named according to the parent reference node.