Complete compile-time ITK v5 migration in All configuration
This is a major milestone of the ITK v5 migration. It is supposed to
result in successful builds of the Differential build clients.
Platform-related fixes may be necessary.

Several tests including critical image statistics tests fail and hence
potential bugs introduced with the migration must be fixed before the
migration is ready for the develop branch.