Improved generic naming and coloring of newly created labels
Names and to a certain extent colors are unique across all layers.
Name gaps like in between "New label 1" and "New label 3" are
intentionally not filled since it indicates that a label with that
name was explicitly deleted before. This mimics the naming behavior of
commonly used software like Photoshop.