Makes picking tool label aware and improves it.
Fixes problems with overwritting other labels.

Now you can also pick from multiple other labels and the picked
elements will be added to the active labels.
Be aware if you pick from other labels the must not be locked,
other wise the lock will be ignored.
If you pick from the active label it is irrelevant if it is locked.