Use UTF-8 encoding on Windows
Added an advanced CMake option MITK_UTF8 (default: ON) to the MITK build.
If enabled, a manifest file is embedded into all MITK applications to
switch to the UTF-8 code page introduced in Windows 10 Version 1903.

It finally solves some issues with characters like German Umlauts in
file names and eases cross-platform development, since both Linux and
macOS are based on UTF-8 as well.