Follow one definition rule
The OrganNamesHandling pseudo struct was included via its definition file
before and hence broke the one definition rule if it was included twice.

It is hard to resist the urge to refactor and fix the whole construct but
for now it is out of scope and was just fixed minimally invasive.