Allow custom label name and color suggestions
Custom label name and color suggestions are feature complete except for
optionally reducing the suggestions by already existing label names in
segmentations.

Since QmitkNewSegmentationDialog needs access to some preferences which
are only available in the context of a plugin, it was moved from the
SegmentationUI module to the segmentation plugin.

Same for QmitkLabelSetWidget, which depends on QmitkNewSegmentationDialog.

All references and dependenceies to these classes were adapted.

QmitkSegmentationOrganNamesHandling was removed as it is no longer needed.