Reduce name property duality
The name property has historically grown into a duality that was never
really resolved. While I/O only has access to base data properties,
renaming in the data manager added/set the name property to/of data nodes,
leading to a drift of two persistent names while there should be only
a single source of truth.

The default signatures of data node property retrival methods query
base data properties as fallback for years now, hence, it is time to
move on and use the base data name property at least if it is already
present. This will hardly break any API but I/O is able to access the
same name as users see in the GUI.