Change type of name control from line edit to combo box
Users can now easily see the complete list of suggestions instead of
having to type something to see only a matching subset.

This change may also be a preparation for later, if we decide to
optionally restrict valid label names to the suggestions, in which case
we would simply switch of the editability of the combobox. Also, SetName()
should then call addItem() with an insert policy instead of setEditText().