Avoid crashes in point picking when props are deleted
I could observe crashes where mappers were deleted and subsequent picking
was then attempting to access some of the deleted vtkProps associated
to the deleted mappers.

I do not see a way to improve the mtime checks in this method to avoid
the aforementioned crash. So the only way to ensure proper picking seems
to be the introduction of a new list of pointers to props so the props
are not getting deleted.

Signed-off-by: Daniel Maleike <daniel.maleike@stryker.com>