Do not always react in PACS mode
The PACS mode would always react, even if an existing data interactor
already handled a specific event. This is not the case for the MITK mode.
In order to be consistent, the always-react-parameter is set to false for
the PACS mode.
Since the PlanarFigureInteractor already handles mouse events we do not
explicitly need to disable / enable the display configuration anymore,
since relevant mouse events are not passed through to the
DisplayActionEventBroadcast-class anymore.