Provide a save button for intermediate results
When the active task is shown, a "Save" button is shown next to the
"Accept segmentation" button. It will just save the current segmentation
as intermediate result instead of saving *and* accepting it.

As soon as an accepted segmentation exists on disk, it will get overridden
instead of creating an intermediate result again.

Intermediate results are stored at the same path as accepted segmentations
but their file extension is prepended by ".intermediate".