Add shortcuts for certain label actions
Label shortcuts are now implemented as key sequences since single key
stroke shortcuts are already using most of the available keys and the
chosen letters tend to be rather obscure at this level.

Label key sequences start with Ctrl+L (for label) and then continue with:
  - N: New label
  - R: Rename label
  - I: Iterate selected label (was Ctrl+L before)

Also fixes the iteration of the selected label since the shortcut allowed
to select the background which is not listed in the label set widget.