Keep last tool active for next segmentation task
The "Accept Segmentation" button already loads the next unfinished task.
We were asked if it would be possible to keep the last selected tool
active as it would greatly simplify the workflow.

If it proves well, we may consider implementing the feature at a more
general location and add a preference for it.