Remove TCP Wrappers (libwrap) dependency
It is optionally (default: on) used by DCMTK for additional access
control cmd-line arguments in storescp and other executables. As
TCP Wrappers is considered deprecated by a few Linux distros like
Fedora and we are not using the optional cmd-line arguments anyway,
we explicitly tell DCMTK to not link against it.

This will probably make our installers compatible to Fedora and
similar distros.