Render window specific geometry comparison
Summary:
Fix broken event definition

Fix return type of the second parameter.

Add new custom event to send two parameters

The new event can send a BaseRenderer and a bool value.
This is used to notify if the given base renderer has been
entered or left with the mouse interaction.

Remove all observer of a tool if destroyed

Modify event filter to include InternalEvents

Provide functions and connect them to internal events

Add transitions and connect them to specific tools

Forward internal event inside QmitkRenderWindow

React on the new custom ToolEnteredRenderWindow-event

This will either add a warning message to the segmentation view
and a warning image to the corresponding base renderer
or remove the warning message and the warning image, depending on the
bool-argument of the event.

Provide function to compare time geometries

This function uses the (time-extracted) world geometry of a
base renderer and compares it with a view-oriented 3D-geometry
of the given time-geometry.

Test Plan:
manual UI testing
1. start MITK, open the segmentation plugin view, load different images
2. create a segmentation node for one of the image
3. activate a tool and try to draw on the render window
	- drawing should be possible but warning signs should appear
	- perhaps you have to perform a global reinit or reint to another node to see the warnings
4. reinit to the reference geometry
	- drawing should be possible without warning signs
5. also test 4D images, static / dynamic segmentation

Reviewers: kislinsk, floca, s434n, O1 MITK Reviewer Group I, O3 MITK Reviewer Group II

Reviewed By: kislinsk, O1 MITK Reviewer Group I, O3 MITK Reviewer Group II

Subscribers: kalali, s434n, floca, kislinsk

Maniphest Tasks: T26376

Differential Revision: https://phabricator.mitk.org/D722