Relax check for empty time steps
While a timestep with a valid polydata but without any content in
that polydata could be considered empty in a practical sense, it
isn't empty in the technical sense, as there actually is a
polydata, even though it is empty.

Time geometry, header, meta and property data is still present and should
be considered when saving such data. Declaring it as empty, though, would
trigger our serialization routines to dismiss all data.