Combine the node predicates for node filtering
The 'fixed layer' node predicate is combined with the user-defined
node predicate member to filter nodes.
This needs to be done either when updating the model (e.g. changing
the data storage or the node predicate) or when a new node is added
to the model by listening to the data storage.