Rename label on ALT + double click
There are two different idioms for double clicking on an item in
applications: Activate/focus/edit the item or rename the item.

Double clicking on a label entry already centers the crosshair to the
centroid of the label. To allow for another quick convenient access
to renaming in addition to "CTRL+L, R", we now also check for the
ALT keyboard modifier on a double click.

Note that it must be ALT since both CTRL and SHIFT are already entangled
in selection interaction like selecting multiple items.