Adding voxel value view to replace status bar information
Summary: This is a first implementation of a new view that simply shows the value of the currently selected voxel in the render windows. Currently it initially shows the wrong value immediately after loading an image. I think it tells the value of the outermost corner. So far I haven't been able to fix this. It also suffers of the same problem as the status bar version, that the values "lag behind" one step when scrolling with the mouse wheel. I have tried forcing a render update, but to no effect. But since this was also the case before, I think it is alright for now.

Test Plan: Open the new "Voxel Value" view, load an image and move around in it. The view should always show the voxel value at the crosshair position. When multiple images are loaded, it should always show the value of the image on top. Should also work for RGB(A) images.

Differential Revision: https://phabricator.mitk.org/D749