React on the GeometryUpdateEvent of the SliceNavigationController
This is required for the mxnMultiWidget since a crosshair reset
will not trigger the already connected GeometrySliceEvent.
Reacting on the GeometrySliceEvent was enough for the StdMultiWidget
since the SliceNavigationController were connected and a crosshair reset
would trigger a slice change in other render windows.
For the mxnMultiWidget this is not the case, so we need to listen to the
GeometryUpdateEvent additionally.