Add cpp-httplib as optional dependency
Switch on MITK_USE_httplib in the MITK superbuild to enable cpp-httplib
in MITK.

We require OpenSSL as prerequisite - if it is not found automatically when
configuring the MITK superbuild, follow the instructions on screen.

To use cpp-httplib in a module or plugin, add "httplib" to its
PACKAGE_DEPENDS list.