Provide functions to enable overlay warnings for not-aligned geometries
In order to move the logic for comparing base renderer and given input
geometry - to find out if a reference geometry aligns with the current
world geometry of a base renderer - additional functions
have been introduced.
SetReferenceGeometry has been renamed to InitializeViews
to better reflect the purpose and to be in line with
similar function names of the RenderingManager.