Fix ImageToContourFilterTest
The filter was changed to accept custom pixel values for
determining the contour. However, this value wasn't initialized
and the calculation for the contour extraction filter was slightly
changed for no obvious reason.

The filter is now fixed to match its previous behaviour when
no custom pixel value is specified.