Fix crash in binary threshold tools
Summary:
Fix crash in binary threshold tools by adapting label set to label mapping.

I don't know if it is preferred but the fix will also show the preview in target label color
instead of default green as before.

This reveals a few minor update issues, e.g., the tools do not react on label color changes or selected label changes,
which would be necessary to always show the preview in the correct color.

Fixes T29575.

Test Plan:
- Create segmentation
- Open "Threshold" or "UL Threshold" tool
- Add another label
- Press "Confirm segmentation"

The tools shouldn't crash anymore and the preview is applied to the selected label.

Reviewers: floca, O1 MITK Reviewer Group I!

Maniphest Tasks: T29575

Differential Revision: https://phabricator.mitk.org/D827