Do not include uic generated file in header
QmitkLevelWindowWidget.h included ui_QmitkLevelWindowWidget.h. This is a file generated by QT's uic. And it is not right to include it in the header. An application that includes QmitkLevelWindowWidget.h fails to compile. A corect way is to do forward declaration in the header and include the generated header in corresponding cpp file. Please see the commit for more details, it is standard way to use uic generated headers, and is already used in many mitk headers already.

Signed-off-by: Sukhraj Singh <sukhraj.singh01@stryker.com>