Support time steps in MITK ROI file format.
The "Geometry" object has an optional "TimeSteps" value specifying
the number of time steps.

The "ROI" object has an optional "TimeSteps" array, which, when present,
contains objects consisting of a mandatory time step number "t",
"Min" and "Max" points, as well as an optional "Properties" object.

"ROI" objects with a "TimeSteps" array are not supposed to contain
"Min" and "Max" points directly anymore. Their "Properties" object
still act as a fallback for the optional "Properties" objects of the
time steps, though.