BREAKING: Refactored LabelSetImage::CreateLabelMask
Why: remove variables useActiveLayer and layer. They are implicitly defined
by index now.
Internally reworked function to simplify code and use multi threading.
How to adapt: Remove the layer variables. The are not needed anymore. If you
used useActiveLayer = false with layers that didn't contain the specified
label (leads to an empty image) this has to be done now by own code
(by checking if the label exists in the specified group).