Add preference page for tool bar visibility
Summary:
See {T30284}.

NOTE: This is not for #mitk-2023-10

Follow the test plan below.

edit: ~~While the preferences are stored at application shutdown, they are not yet applied at application startup. I have not yet looked into how to do this without interfering with the Workbench Window Advisors. This will be the next iteration of this Diff if it is decided that the direction of it is okay in general.~~

Fixes T30284.

Test Plan:
- Open MITK Workbench or MITK Flowbench
- Open Preferences (Ctrl+P) and navigate to the "Tool Bars" page / type "tool" in the filter.
- Check if all tool bars are listed, checked, and collapsed
- Uncheck a few tool bars, apply the preference changes and check if these tool bars are now hidden in the application
- Make a few of them visible again in the preferences and check again
- Exit and start the application again and check that the tool bars preferences are still/already applied

Reviewers: floca, O1 MITK Reviewer Group I

Reviewed By: floca, O1 MITK Reviewer Group I

Maniphest Tasks: T30284

Differential Revision: https://phabricator.mitk.org/D890