BREAKING: DICOMGDCMImageFrameInfo does not return a  default image position or orientation any more.
Why: Lead to inconsistencies/exceptions in other reader code
because it was not clear if data was missing or realy that values.
That default branch is very old and was coded before
DICOMDatasetFinding were introduced. The allow now to communicate
if a value is valid or not. This should be used by calling
code to decide how to behave if tags are not present.