BREAKING CHANGE: mitk::MultiLabelSegmentation::AddLabel() does not change active label anymore.
Why: Implicit changing of active label lead in some cases to invalid states
of MultiLabelSegmentationsInspectors, where selection in UI was not in sync
with behavior. As the ActiveLabel feature is anyway deprecated, it was
decided to make this change of label explicit where needed and avoid doing
the change when adding a label.

How to adapt: If you need the active label set after the label adding, do
it explicitly via MultiLabelSegmentation::SetActiveLabel(...). See also
changes in this commit for examples.