Change default build type to Release and warn on Debug Python builds
Release as default build type is more common in projects as it usually
is the desired build type except for specific developer needs, in which
case we can assume a developer is able to switch the build type.

Debug builds with Python are troublesome and should be avoided in general.
Hence, we warn about or remove Debug from the list of suggested
configuration types as soon as MITK_USE_Python3 is enabled.