Fix interference between WorkbenchRelease and WorkbenchReleaseNoPython
Every second CMake configure run, SWIG and Python variables were removed
from the cache because the WorkbenchRelease configuration included
the WorkbenchReleaseNoPython configuration.

Both configurations now share a common base file to prevent mutual
interference.

The base file is placed in a subfolder to exclude it from the
auto-detection of selectable configurations for users.