Use Doxygen variable for required OpenSSL version
The required version of OpenSSL on Windows is now defined by
MITK_REQUIRED_OPENSSL_VERSION.

We no longer search or recommend OpenSSL as provided by the
Qt Installer. It was recently removed from the installer, and
older installer versions may not provide an OpenSSL version
that is compatible with our Python dependency.